Create any classes

+++ src/api/MedicalClinic.Domain/Entities/Insurance.cs
@@ -14,10 +14,9 @@ namespace MedicalClinic.Domain.Entities
         public string Name { get; set; } = null!;
         public string? Description { get; set; }
 
-        //TODO: To Alter Propoerties to double
-        public decimal PercentageTypeOne { get; set; }
-        public decimal PercentageTypeTwo { get; set; }
-        public decimal PercentageTypeThree { get; set; }
+        public double PercentageTypeOne { get; set; }
+        public double PercentageTypeTwo { get; set; }
+        public double PercentageTypeThree { get; set; }
 
         public bool IsEnabled { get; set; }
         public DateTime? DeletedAt { get; set; }
